Separeted CalendarInline in View and ViewModel
